Refactors datasets
This commit refactors datasets management to accommodate to datasets with
provided train/val/test splits.

Previous datasets are managed using the Dataset class that is responsible
of creating the test partition and the outer CV splits with the specified
parameter.

Datasets with provided train/val/test splits are managed using the
PartitionedDataset class that returns a single split in the outer CV using
train/val.

This allows reusing the existing nested-CV code with these new type of
datasets.